Migrated asset pipeline from Sprockets to Propshaft
