--- app/assets/config/manifest.js
@@ -1,9 +1,0 @@
-//= link_tree ../javascripts .js
-
-//= link_tree ../../../vendor/assets/javascripts .js
-//= link cookies_eu.js
-
-// Images and fonts so that views can link to them
-//
-//= link_tree ../images
-//= link_tree ../builds
