+++ package.json
@@ -50,9 +50,11 @@
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^14.6.1",
+    "autoprefixer": "^10.5.0",
     "babel-jest": "^30.0.5",
     "babel-loader": "^10.0.0",
     "babel-plugin-prismjs": "^2.1.0",
+    "copy-webpack-plugin": "^14.0.0",
     "css-loader": "^7.1.2",
     "jest": "^30.0.5",
     "jest-canvas-mock": "^2.5.2",
@@ -60,6 +62,8 @@
     "jest-fetch-mock": "^3.0.3",
     "jsdom-worker": "^0.3.0",
     "mini-css-extract-plugin": "^2.9.2",
+    "postcss": "^8.5.15",
+    "postcss-loader": "^8.2.1",
     "sass": "^1.89.2",
     "sass-loader": "^16.0.5",
     "webpack": "^5.105.0",
@@ -72,8 +76,6 @@
   "scripts": {
     "build": "node_modules/.bin/webpack --progress --config webpack.production.js",
     "build-dev": "node_modules/.bin/webpack --progress --config webpack.development.js",
-    "build:css": "node_modules/.bin/webpack --progress --config webpack.production.js",
-    "build-dev:css": "node_modules/.bin/webpack --progress --config webpack.development.js",
     "test": "node_modules/.bin/jest --no-cache",
     "test-cov": "node_modules/.bin/jest --no-cache --collect-coverage"
   },
